+++ src/style.rs
@@ -6,7 +6,7 @@ pub enum StyleVar {
     WindowPadding(ImVec2),
     WindowRounding(f32),
     WindowMinSize(ImVec2),
-    ChildWindowRounding(f32),
+    ChildRounding(f32),
     FramePadding(ImVec2),
     FrameRounding(f32),
     ItemSpacing(ImVec2),
@@ -14,4 +14,12 @@ pub enum StyleVar {
     IndentSpacing(f32),
     GrabMinSize(f32),
     ButtonTextAlign(ImVec2),
+}
+
+impl StyleVar {
+    #[allow(non_snake_case)]
+    #[deprecated(since = "0.0.19", note = "please use ChildRounding instead")]
+    pub fn ChildWindowRounding(value: f32) -> Self {
+        StyleVar::ChildRounding(value)
+    }
 }